op-chain-ops: more cleanup
